% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryNames.R
\docType{data}
\name{CountryNames}
\alias{CountryNames}
\title{Country Names}
\format{
A \code{vector} with 1627 entries.
}
\usage{
CountryNames
}
\description{
A \code{vector} that has all the names of countries around the world
}
\keyword{datasets}
