% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHeader.R
\name{CreateHeader}
\alias{CreateHeader}
\title{CreateHeader}
\arguments{
\item{x}{a \code{dataframe}.}
}
\value{
'CreateHeader()' returns a \code{dataframe} where the first row is now the header
}
\description{
Takes the first row a \code{dataframe} and makes it the header.
}
\details{
'CreateHeader()' takes the first row a \code{dataframe} and makes it the header.
}
\examples{
df <- data.frame(x = c("First", "Sam"), y = c("Height", 178))
CreateHeader(df)
}
