% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftTeam.R
\name{DraftTeam}
\alias{DraftTeam}
\title{DraftTeam}
\arguments{
\item{x}{a \code{dataframe}.}
}
\value{
'DraftTeam()' returns a \code{dataframe} with both the draft year and team that drafted a player.
}
\description{
Takes the first row a \code{dataframe} and makes it the header.
}
\details{
'DraftTeam()' is used to extract the draft year and draft team from a player profile
}
