% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuroScrapR.R
\name{EuroScrapR}
\alias{EuroScrapR}
\title{EuroScrapR}
\arguments{
\item{Season}{List of two year ranges for season}

\item{Country}{Character of countries}

\item{League}{Character of leagues}

\item{Tournament}{Character of tournament}

\item{Team}{Character of team}

\item{PlayerLastName}{Character last name of player}

\item{PlayerFirstName}{Character first name of player}

\item{Position}{Character position}
}
\value{
'EuroScrapR()' returns a \code{dataframe} containing the desired data from EuroBasket.com
}
\description{
Allows users to select data collected from EuroBasket.com
}
\details{
'EuroScrapR()' takes inputs from the users to select only the desired data from Eurobasket.com.
Some of these inputs include Season, Country, League, Tournament, Team, Last Name of Player, First Name of Player, and Position.
All of these inputs are optional.
}
\examples{
EuroScrapR(Position = "Point Guard")
EuroScrapR(Season = "2018-2019")
EuroScrapR(Season = c("2017-2018","2018-2019"))
}
