% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerHrefGrab.R
\name{PlayerHrefGrab}
\alias{PlayerHrefGrab}
\title{PlayerHrefGrab}
\usage{
PlayerHrefGrab(URL)
}
\arguments{
\item{URL}{a URL taken from one of the stats pages on EuroBasket.com}
}
\value{
'PlayerHrefGrab()' returns a \code{dataframe} with the player's hrefs, team href, and player ids.
}
\description{
Allows the user to input a stats URL from EuroBasket.com and collects all the player hrefs
 on the page.
}
\details{
'PlayerHrefGrab' takes a url taken from one of the stats pages on Eurobasket.com and returns
a \code{dataframe} with all the player hrefs, team hrefs, and player ids.
}
\examples{
PlayerHrefGrab(URL = "https://basketball.eurobasket.com/team/Spain/BAXI-Manresa/318?Page=3")
}
