% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerInfoGrab.R
\name{PlayerInfoGrab}
\alias{PlayerInfoGrab}
\title{PlayerInfoGrab}
\usage{
PlayerInfoGrab(df)
}
\arguments{
\item{df}{a \code{dataframe} with the players' profile URL in the first collumn with the team URL in the second collumn
and the player id in the third collumn. The second and third collumn are optional but it must but the input must be a
\code{dataframe}.}
}
\value{
'PlayerInfoGrab()' returns a \code{dataframe} with the players' info.
}
\description{
Takes a player profile URL from EuroBasket.com and returns a \code{dataframe} with the players info.
}
\details{
'PlayerInfoGrab()' Takes a player profile URL from EuroBasket.com and returns a \code{dataframe}
 with the players info.
}
\examples{
df <- data.frame("https://basketball.asia-basket.com/player/Jimmer-Fredette/China/Shanghai-Sharks/135233",
                  "https://basketball.asia-basket.com/team/China/Shanghai-Sharks/1950?Page=3",
                  "135233")

PlayerInfoGrab(df)

df1 <- data.frame("https://basketball.asia-basket.com/player/Jimmer-Fredette/China/Shanghai-Sharks/135233",
                  "a",
                  "a")

PlayerInfoGrab(df1)
}
