% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonStatGrab.R
\name{SeasonStatGrab}
\alias{SeasonStatGrab}
\title{SeasonStatGrab}
\usage{
SeasonStatGrab(URL)
}
\arguments{
\item{URL}{a URL taken from one of the stats pages on EuroBasket.com}
}
\value{
'SeasonStatGrab()' returns a \code{dataframe} containing data from one of the EuroBasket.com stat pages
}
\description{
Allows the user to input a url from Eurobasket and scrap the data from the page.
}
\details{
'SeasonStatGrab' takes a url taken from one of the stats pages on Eurobasket.com and returns
a \code{dataframe} with the sraped data.
}
\examples{
SeasonStatGrab(URL = "https://basketball.eurobasket.com/team/Spain/BAXI-Manresa/318?Page=3")
}
