% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_rebound.R
\name{rename_rebound}
\alias{rename_rebound}
\title{rename_rebound}
\usage{
rename_rebound(x)
}
\arguments{
\item{x}{a \code{dataframe} given while running the SeasonStatGrab() function}
}
\value{
'rename_rebound()' retuns a \code{dataframe} in which all the abbreviations for rebound have been
given a uniform name
}
\description{
This is a function used in SeasonStatGrab() that changes the abbreviations of rebounds
from "OFF" to "RO", "DEF" to "RD", and "TOT" to "REB"
}
\details{
'rename_rebound' takes a \code{dataframe} while in the function SeasonStatGrab() and
changes the abbreviatons for offensive, defensive, and total rebounds to a uniform "RO", "RD", and "REB"
}
